# metal : fix fa kernel