@slaren 2024-08-26
I cannot reproduce this on M3 Max, but a race condition seems like it would be the most likely cause. If that's the case, it is likely to create more issues in the future, so it would be good to fix it completely rather than making some changes that make the issue less likely, but still possible.