@ggerganov 2024-08-26
I found another fix with #9188, but I still don't understand why `master` fails. It also fails on M1 Pro. Additionally, every generation with `llama-cli` and a fixed random seed results in a different output, which does indicate a race condition.

Another observation is that the race is likely inside this block:

https://github.com/ggerganov/llama.cpp/blob/7a3df798fc43e1b366146b1ad99137a9e95c87dd/ggml/src/ggml-metal.metal#L2142-L2160

The reason is that if I change the `tx` and `ty` vars to be loops in `[0..4)` and `[0..8)` respectively, using only `tiisg == 0`, the results are again correct:

```diff
diff --git a/ggml/src/ggml-metal.metal b/ggml/src/ggml-metal.metal
index aba0b9a0..df1be087 100644
--- a/ggml/src/ggml-metal.metal
+++ b/ggml/src/ggml-metal.metal
@@ -2141,8 +2141,9 @@ kernel void kernel_flash_attn_ext_f16(
 
                     simdgroup_store(mqk, ss + 8*cc, TF, 0, false);
 
-                    const short tx = tiisg%4;
-                    const short ty = tiisg/4;
+                    if (tiisg == 0) {
+                        for (short ty = 0; ty < 8; ++ty) {
+                            for (short tx = 0; tx < 4; ++tx) {
 
                     // mqk = mqk*scale
                     ss[8*cc + ty*TF + 2*tx + 0] *= scale;
@@ -2157,6 +2158,10 @@ kernel void kernel_flash_attn_ext_f16(
                         // mqk = mqk + mask*slope
                         ss[8*cc + ty*TF + 2*tx + 0] += slope*mp[ic + 8*cc + ty*nb31/sizeof(half) + 2*tx + 0];
                         ss[8*cc + ty*TF + 2*tx + 1] += slope*mp[ic + 8*cc + ty*nb31/sizeof(half) + 2*tx + 1];
+                    }
+
+                            }
+                        }
                     }
                 }
             }
```

Any ideas what could be the issue? I'm not sure how to debug this further