@slaren 2024-08-26
I don't know what threads use what data, and I cant' test it either. Intuitively I would assume that a barrier is necessary before the softmax to synchronize the shared memory in `ss`.