@ggerganov 2024-08-26
It does look like a barrier is necessary, though adding one did not fix the problem.

However, I rewrote the code to make the threads access the data more coherently and this resolved the problem: https://github.com/ggerganov/llama.cpp/pull/9189